# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PamliEcoConnect REST API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List all team members
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team, admin]
      summary: Create a team member (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberBody"
      responses:
        "201":
          description: Created team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team, admin]
      summary: Update a team member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team, admin]
      summary: Delete a team member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Contact
...
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        headshotUrl:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - title
        - bio
        - displayOrder
        - createdAt
    TeamMemberBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        headshotUrl:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - name
        - title
        - bio

    # Contact
...